+++ packages/bsync/src/client.ts
@@ -3,7 +3,7 @@ import {
   ConnectTransportOptions,
   createConnectTransport,
 } from '@connectrpc/connect-node'
-import { Service } from './gen/bsky_sync_connect'
+import { Service } from './gen/bsync_connect'
 
 export type BsyncClient = PromiseClient<typeof Service>
