+++ packages/bsync/src/db/schema/mute_op.ts
@@ -1,5 +1,5 @@
 import { GeneratedAlways, Selectable } from 'kysely'
-import { MuteOperation_Type } from '../../gen/bsky_sync_pb'
+import { MuteOperation_Type } from '../../gen/bsync_pb'
 
 export interface MuteOp {
   id: GeneratedAlways<number>
