+++ packages/bsync/src/gen/bsync_connect.ts
@@ -1,9 +1,9 @@
 // @generated by protoc-gen-connect-es v1.3.0 with parameter "target=ts,import_extension=.ts"
-// @generated from file bsky_sync.proto (package bsync, syntax proto3)
+// @generated from file bsync.proto (package bsync, syntax proto3)
 /* eslint-disable */
 // @ts-nocheck
 
-import { AddMuteOperationRequest, AddMuteOperationResponse, PingRequest, PingResponse, ScanMuteOperationsRequest, ScanMuteOperationsResponse } from "./bsky_sync_pb.ts";
+import { AddMuteOperationRequest, AddMuteOperationResponse, PingRequest, PingResponse, ScanMuteOperationsRequest, ScanMuteOperationsResponse } from "./bsync_pb.ts";
 import { MethodKind } from "@bufbuild/protobuf";
 
 /**
