+++ packages/bsync/src/gen/bsync_pb.ts
@@ -1,5 +1,5 @@
 // @generated by protoc-gen-es v1.6.0 with parameter "target=ts,import_extension=.ts"
-// @generated from file bsky_sync.proto (package bsync, syntax proto3)
+// @generated from file bsync.proto (package bsync, syntax proto3)
 /* eslint-disable */
 // @ts-nocheck
